componaize
